Refactor to make testing easier
