+++ src/lib/importSystem/process.ts
@@ -1,16 +1,4 @@
-import { Op } from 'sequelize';
-import { Readable } from 'stream';
-import {
-  remap,
-  collectChangedValues,
-  lowercaseKeys,
-  createRanges,
-} from '../dataObjectUtils';
-import { filterDataToModel, modelForTable } from '../modelUtils';
-import EncodingConverter from '../stream/encoding';
-import Hasher, { getHash } from '../stream/hasher';
 import S3Client from '../stream/s3';
-import XMLStream, { SchemaNode } from '../stream/xml';
 import ZipStream, { FileInfo as ZipFileInfo } from '../stream/zip';
 import {
   getNextFileToProcess,
@@ -20,322 +8,9 @@ import {
   updateAvailableDataFileMetadata,
 } from './record';
 import { IMPORT_DATA_STATUSES, IMPORT_STATUSES } from '../../constants';
-import db from '../../models';
 import { auditLogger } from '../../logger';
-
-type ProcessDefinition = {
-  fileName: string,
-  encoding: string,
-  tableName: string,
-  keys: string[],
-  remapDef: Record<string, string>;
-};
-
-/**
- * Process records according to the given process definition and XML client.
- * @param processDefinition - The process definition object.
- * @param xmlClient - The XML client object.
- * @param fileDate -  the data the file was modified
- * @param recordActions - The record actions object containing arrays of promises for
- * inserts, updates, and deletes.
- * @param schema - the name of each of the columns within the data
- * @returns A promise that resolves to the updated recordActions object and schema.
- */
-const processRecords = async (
-  processDefinition: ProcessDefinition,
-  xmlClient: XMLStream,
-  fileDate: Date,
-  recordActions: {
-    inserts,
-    updates,
-    deletes,
-    errors,
-  } = {
-    inserts: [],
-    updates: [],
-    deletes: [],
-    errors: [],
-  },
-): Promise<{
-  inserts,
-  updates,
-  deletes,
-  errors,
-}> => {
-  let record;
-  try {
-    record = await xmlClient.getNextObject(true);
-  } catch (err) {
-    // record the error into the recordActions and continue on successfully as
-    // other entries may be process successfully
-    recordActions.errors.push(err.message);
-    auditLogger.log('error', ` processRecords getNextObject ${err.message}`, err);
-  }
-
-  // @ts-ignore
-  let model;
-  try {
-    model = modelForTable(db, processDefinition.tableName);
-  } catch (err) {
-    // record the error into the recordActions
-    recordActions.errors.push(err.message);
-    auditLogger.log('error', ` processRecords modelForTable ${err.message}`, err);
-
-    // Unable to continue as a model is required to record any information
-    return Promise.reject(recordActions);
-  }
-
-  if (record) {
-    try {
-      // TODO: column/key alpha sort to retain order
-      // 1. use the remap method to format data to structure needed
-      // 2. use the filterDataToModel to match what is expected
-      // 3. check for existing record
-      // 4a. if new
-      //   1. insert
-      //   2. recordActions.inserts.push(uuid)
-      // 4b. if found
-      //   1. use the collectChangedValues to find the values to update
-      //   2. update
-      //   2. recordActions.update.push(uuid)
-
-      // Format the record data using the remap method
-      // This changes the attribute names and structure into what will be saved
-      const { mapped: data } = remap(
-        record,
-        lowercaseKeys(processDefinition.remapDef),
-        {
-          keepUnmappedValues: false,
-          // defines a custom fuction that will replace the resulting structure
-          // with the result of each function.
-          targetFunctions: {
-            // take in an object and generate a hash of that object
-            'toHash.*': (toHash) => ({ hash: getHash(toHash) }),
-          },
-        },
-      );
-
-      // Filter the data to match the expected model
-      const {
-        matched: filteredData,
-        unmatched: droppedData,
-      } = await filterDataToModel(data, model);
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const recordKey: Record<string, any> = {};
-      processDefinition.keys.forEach((key) => {
-        const value = filteredData[key];
-        if (value) {
-          recordKey[key] = value;
-        }
-        // TODO: handle case where all/part of the key may have been dropped
-      });
-      if (Object.keys(droppedData).length > 0) {
-        // TODO: add some kind of note/warning that mapped data was filtered out at the model level
-        // The message should include the importDataFileId, the recordKey, and the column names.
-        // The column values should be excluded to prevent posable display of PII
-      }
-
-      // Check if there is an existing record with the same key value
-      const currentData = await model.findOne({
-        where: {
-          ...recordKey,
-        },
-      });
-
-      if (!currentData) {
-        // If the record is new, create it
-        const insert = model.create(
-          {
-            ...filteredData,
-            sourceCreatedAt: fileDate,
-            sourceUpdatedAt: fileDate,
-          },
-          {
-            individualHooks: true,
-            returning: true,
-          },
-        );
-        recordActions.inserts.push(insert);
-      } else if (fileDate > currentData.sourceUpdatedAt) {
-        // If the record already exists, find the delta then update it
-        const delta = collectChangedValues(filteredData, currentData);
-        const update = model.update(
-          {
-            ...delta,
-            sourceUpdatedAt: fileDate,
-            ...(currentData.sourceDeletedAt && { sourceDeletedAt: null }),
-            updatedAt: new Date(),
-          },
-          {
-            where: { id: currentData.id },
-            individualHooks: true,
-            returning: true,
-          },
-        );
-        recordActions.updates.push(update);
-      }
-    } catch (err) {
-      // record the error into the recordActions and continue on successfully as
-      // other entries may be process successfully
-      recordActions.errors.push(err.message);
-      auditLogger.log('error', ` processRecords create/update ${err.message}`, err);
-    }
-  } else {
-    try {
-      // 1. Find all records not in recordActions.inserts and recordActions.update
-      // 2. delete
-      // 3. recordActions.delete.push(promises)
-      // 4. pass back recordActions
-
-      const [
-        affectedDataInserts,
-        affectedDataUpdates,
-      ] = await Promise.all([
-        Promise.all(recordActions.inserts),
-        Promise.all(recordActions.updates),
-      ]);
-
-      // Flatten the affectedDataUpdates array and extract the objects
-      const flattenedUpdates = affectedDataUpdates.flatMap(
-        // Assuming the second element of each sub-array is the array of objects
-        (update) => (Array.isArray(update[1]) ? update[1] : []),
-      );
-
-      // Combine the affected data from inserts and flattened updates
-      const affectedData = [
-        ...affectedDataInserts,
-        ...flattenedUpdates,
-      ];
-
-      const affectedDataIds = affectedData?.map(({ id }) => id).filter((id) => id) || [];
-      const affectedRanges = createRanges(affectedDataIds);
-
-      // mark the source date when the records no longer are present in the processed file
-      // "Delete" all records that are not in the affectedData array
-      if (affectedDataIds.length) {
-        const destroys = model.update(
-          {
-            sourceDeletedAt: fileDate,
-          },
-          {
-            where: {
-              [Op.and]: affectedRanges.map((range) => ({
-                id: { [Op.notBetween]: range },
-              })),
-              sourceDeletedAt: null,
-            },
-            individualHooks: true,
-          },
-        );
-
-        recordActions.deletes.push(destroys);
-      }
-    } catch (err) {
-      // record the error into the recordActions
-      recordActions.deletes.push(err.message);
-      auditLogger.log('error', ` processRecords destroy ${err.message}`, err);
-    }
-
-    return Promise.resolve(recordActions);
-  }
-
-  // Recursively call the processRecords function to process the next record
-  return processRecords(
-    processDefinition,
-    xmlClient,
-    fileDate,
-    recordActions,
-  );
-};
-
-/**
- * Processes a file based on the provided process definition.
- *
- * @param processDefinition - The process definition object that contains information
- * about how to process the file.
- * @param fileInfo - Information about the file being processed.
- * @param fileStream - The stream of the file being processed.
- * @returns A promise that resolves to an object containing arrays of promises for
- * inserts, updates, and deletes.
- * @throws An error if the remapDefs property is not found in the processDefinition.
- * @throws An error if the model property is not found in the processDefinition.
- * @throws An error if the key property is not found in the processDefinition.
- */
-const processFile = async (
-  processDefinition: ProcessDefinition,
-  fileInfo: ZipFileInfo,
-  fileStream: Readable,
-): Promise<{
-  hash?: string,
-  schema?: SchemaNode,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  inserts?: Promise<any>[],
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  updates?: Promise<any>[],
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  deletes?: Promise<any>[],
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  errors: Promise<any>[],
-}> => {
-  let result: {
-    hash?: string,
-    schema?: SchemaNode,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    inserts?: Promise<any>[],
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    updates?: Promise<any>[],
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    deletes?: Promise<any>[],
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    errors: Promise<any>[],
-  } = {
-    errors: [],
-  };
-
-  try {
-    // Check if remapDefs property exists in processDefinition, if not throw an error
-    if (!processDefinition?.remapDef) throw new Error('Remapping definitions not found');
-    // Check if model property exists in processDefinition, if not throw an error
-    if (!processDefinition?.tableName) throw new Error('Model not found');
-    // Check if key property exists in processDefinition, if not throw an error
-    if (!processDefinition?.keys) throw new Error('Keys not found');
-    // Check if key property exists in processDefinition, if not throw an error
-    if (!processDefinition?.encoding) throw new Error('Encoding not found');
-
-    const hashStream = new Hasher('sha256');
-
-    const encodingConverter = new EncodingConverter('utf8', processDefinition.encoding);
-
-    // Convert the fileStream to a usable stream while also calculation the hash
-    const usableStream = fileStream.pipe(hashStream).pipe(encodingConverter);
-
-    // Create a new instance of XMLStream using the usableStream
-    const xmlClient = new XMLStream(usableStream, true);
-    await xmlClient.initialize();
-
-    // Check if key property exists in processDefinition, if not throw an error
-    if (!xmlClient) throw new Error('XMLStream failed');
-
-    const processedRecords = await processRecords(processDefinition, xmlClient, fileInfo.date);
-
-    // hash needs to be collected after processRecords returns to make sure all the data has
-    // been processed for all records in the file
-    const hash = await hashStream.getHash();
-    const schema = await xmlClient.getObjectSchema();
-
-    result = {
-      hash,
-      schema,
-      ...processedRecords,
-    };
-  } catch (err) {
-    result.errors.push(err.message);
-    auditLogger.log('error', ` processFile ${err.message}`, err);
-  }
-
-  return result;
-};
+import processFile from './processFile';
+import { ProcessDefinition } from './types';
 
 /**
  * Processes the files using the provided ZipStream object and the array of files to process.
@@ -598,8 +273,6 @@ const processZipFileFromS3 = async (
 };
 
 export {
-  processRecords,
-  processFile,
   processFilesFromZip,
   processZipFileFromS3,
 };
