+++ src/lib/importSystem/types.ts
@@ -1,0 +1,7 @@
+export type ProcessDefinition = {
+  fileName: string,
+  encoding: string,
+  tableName: string,
+  keys: string[],
+  remapDef: Record<string, string>;
+};
